+++ pi-monitor/pi-monitor-client/browser/react/src/jvmTest/kotlin/acceptance/utils/AcceptanceTest.kt
@@ -2,9 +2,9 @@ package acceptance.utils
 
 import pimonitor.Application
 import pimonitor.WebApplication
-import testing.ContainerTest
+import testing.containers.ContainerTest.Companion.urlUnderTest
 
-open class AcceptanceTest : ContainerTest() {
+open class AcceptanceTest : ContainerTest {
     companion object {
         val application: Application get() = WebApplication(urlUnderTest)
     }