+++ pi-monitor/pi-monitor-test/containers/src/main/kotlin/testing/containers/RootProjectDir.kt
@@ -1,7 +1,6 @@
-package testing
+package testing.containers
 
 import kotlin.io.path.Path
-import kotlin.io.path.absolute
 import kotlin.io.path.absolutePathString
 
 class RootProjectDir {