+++ pi-monitor/pi-monitor-test/testing/src/commonMain/kotlin/testing/ContainerTest.kt
@@ -1,0 +1,8 @@
+package testing
+
+expect open class ContainerTest() {
+    companion object {
+        val mode: TestMode
+        val urlUnderTest: String
+    }
+}