+++ pi-monitor/pi-monitor-test/testing/src/commonMain/kotlin/testing/annotations/Testcontainers.kt
@@ -1,0 +1,5 @@
+package testing.annotations
+
+@OptIn(ExperimentalMultiplatform::class)
+@OptionalExpectation
+expect annotation class Testcontainers()