+++ pi-monitor/pi-monitor-test/testing/src/jsMain/kotlin/testing/IntegrationTest.kt
@@ -1,0 +1,7 @@
+package testing
+
+import bitframe.service.config.ServiceConfig
+
+actual open class IntegrationTest : ContainerTest() {
+    actual val config: ServiceConfig = ServiceConfig.DEFAULT
+}