+++ pi-monitor/pi-monitor-test/testing/src/jvmMain/kotlin/testing/annotations/Testcontainers.kt
@@ -1,0 +1,5 @@
+package testing.annotations
+
+import org.testcontainers.junit.jupiter.Testcontainers
+
+actual typealias Testcontainers = Testcontainers