+++ CHANGELOG.md
@@ -9,6 +9,11 @@
 ### Bug Fixes
 
 - [[BF92]](https://github.com/picortex/bitframe/issues/92) Fixed sign in validation bug for empty sign in credentials
+- [[BF93]](https://github.com/picortex/bitframe/issues/93) Attach Universal testing to cover future bugs
+
+## Bitframe Test Containers
+
+- [[BF96]](https://github.com/picortex/bitframe/issues/96) Make PiMonitor bitframe containers multiplatform
 
 # 0.0.28